requirements.txt
# FAIR-Agent Requirements
# Dependencies for the FAIR-Agent multi-agent system

# Core ML and NLP libraries
torch>=1.9.0
transformers>=4.20.0
datasets>=2.0.0
accelerate>=0.20.0

# Language Model frameworks
langchain>=0.0.200
openai>=0.27.0  # For potential GPT integration
anthropic>=0.3.0  # For potential Claude integration

# Data processing and analysis
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=0.19.0
click>=8.0.0  # For CLI interface
tqdm>=4.62.0  # Progress bars

# Evaluation and metrics
rouge-score>=0.1.2
bert-score>=0.3.13
bleurt>=0.0.2
nltk>=3.7

# Medical domain specific
bio-embeddings>=0.2.2
scispacy>=0.5.0
medspacy>=1.0.0

# Finance domain specific
yfinance>=0.1.70  # For financial data
quantlib>=1.26  # Quantitative finance

# Web and API frameworks (for future API development)
fastapi>=0.68.0
uvicorn>=0.15.0
requests>=2.26.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.0.0

# Testing
pytest>=6.2.0
pytest-cov>=2.12.0
pytest-mock>=3.6.0

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=0.5.0

# Development tools
black>=21.0.0  # Code formatting
flake8>=3.9.0  # Linting
isort>=5.9.0  # Import sorting
pre-commit>=2.15.0  # Git hooks

# Logging and monitoring
loguru>=0.6.0
wandb>=0.12.0  # Experiment tracking
tensorboard>=2.7.0

# Database (for caching and storage)
sqlite3  # Built-in Python
redis>=3.5.0  # Optional for advanced caching

# Security
cryptography>=3.4.0

# Environment management
python-decouple>=3.4

# Optional GPU acceleration
# Uncomment if using CUDA
# torch-audio>=0.9.0
# torchvision>=0.10.0

# Development dependencies (install with pip install -e ".[dev]")
[dev]
jupyter>=1.0.0
ipython>=7.25.0
notebook>=6.4.0